backend/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
httpx==0.28.1
python-dotenv==1.2.1
openai==2.8.0
google-generativeai==0.8.3
pydantic==2.12.4
python-multipart==0.0.12
